selftests: remove udev rules affecting
Certain distributions have hacks in place to prevent
the firmware loader fallback mechanism from actually
being used. Debian is one of them, a udev rule is set
in place in place to immediately cancel *all* fallback
requests. This was done given that only a few drivers
exist which *really* want the fallback mechansim to be
used on purpose without a timeout, and if the fallback
mechanism is used but finds no file present, it can run
forever stalling the kernel driver making this request.

This udev rule is a work around to buggy drivers which
request no timeout by mistake. But it also means these
distributions using this udev rule cannot every allow
the fallback mechanism. Only two known drivers currently
rely on the fallback mechanism on purpose, but they are
oddball drivers:

 * dell_rbu
 * leds-lp55xx

This fails the firmware API selftest test with:

===========================================
  ./tools/testing/selftests/firmware/fw_fallback.sh: fallback mechanism immediately cancelled

  The file never appeared: /sys/devices/virtual/misc/test_firmware/nope-test-firmware.bin/loading
===========================================

Debian uses the udev fule file /lib/udev/rules.d/50-firmware.rules
with this in it:

SUBSYSTEM=="firmware", ACTION=="add", ATTR{loading}="-1"

Kill this file if present. Although this is known to be a debian
specific work around just ensure to delete this file for all
distributions just in case, and allow us to easily grow removal
of these files by only needing to extend this role with one line
if other files are known to cause issues.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>